Separate serialize functions into separate file
