+++ src/main/scala/firrtl/WIR.scala
@@ -29,6 +29,7 @@ package firrtl
 
 import scala.collection.Seq
 import Utils._
+import firrtl.Serialize._
 import WrappedExpression._
 import WrappedWidth._
