+++ src/main/scala/firrtl/passes/Passes.scala
@@ -40,6 +40,7 @@ import scala.collection.mutable.ArrayBuffer
 
 import firrtl._
 import firrtl.Utils._
+import firrtl.Serialize._
 import firrtl.PrimOps._
 import firrtl.WrappedExpression._
